Added permission check for deleting item.

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -36,7 +36,14 @@ const Mutations = {
   },
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
-    const item = await ctx.db.query.item({ where }, '{ title id description }');
+    const item = await ctx.db.query.item({ where }, '{ title id user { id } }');
+
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission => ['ADMIN', 'ITEMDELETE'].includes(permission));
+
+    if (!ownsItem && !hasPermissions) {
+      throw new Error('You don\'t have permission to do that.');
+    }
 
     return ctx.db.mutation.deleteItem({ where }, info);
   },
